size of string

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -7,7 +7,7 @@
  * Return: returns the numbers of caracters in a string
 */
 
-size_t  _strlen(char *s)
+int  _strlen(char *s)
 {
 	size_t  size;
